+++ spec/internal/app/models/person.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
 class Person < ApplicationRecord
-	belongs_to :parent, class_name: 'Person'
+	belongs_to :parent, class_name: 'Person', inverse_of: :children
+	has_many :children, class_name: 'Person', inverse_of: :parent
 end